+++ lib/td_bg_web/views/business_concept_version_view.ex
@@ -118,6 +118,7 @@ defmodule TdBgWeb.BusinessConceptVersionView do
       business_concept_id: business_concept_version.business_concept.id,
       type: business_concept_version.business_concept.type,
       content: business_concept_version.content,
+      confidential: business_concept_version.business_concept.confidential,
       completeness: Map.get(business_concept_version, :completeness),
       name: business_concept_version.name,
       description: business_concept_version.description,